requirements.txt
# MCP Xcode Server Dependencies
# ================================
#
# Compatible with Python 3.10+
#

# MCP Protocol SDK
mcp>=1.0.0

# Ollama Client
ollama>=0.3.0

# Vector Database (LanceDB - Python 3.14 compatible)
lancedb>=0.6.0
pyarrow>=15.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Async support
asyncio-atexit>=1.0.1
httpx>=0.27.0

# Logging and debugging
rich>=13.7.0
loguru>=0.7.2

# File processing
watchdog>=4.0.0
gitignore-parser>=0.1.11

# Development/Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
